Update CMake to work with Intel compilers

+++ compress/directxtest.cpp
@@ -48,10 +48,10 @@ bool RunTests()
 
     _controlfp_s(nullptr, 0, _EM_ZERODIVIDE | _EM_OVERFLOW | _EM_INVALID);
 
-    UINT nPass = 0;
-    UINT nFail = 0;
+    size_t nPass = 0;
+    size_t nFail = 0;
 
-    for( UINT i=0; i < ( sizeof(g_Tests) / sizeof(TestInfo) ); ++i)
+    for(size_t i=0; i < ( sizeof(g_Tests) / sizeof(TestInfo) ); ++i)
     {
         print("%s: ", g_Tests[i].name );
 
@@ -85,7 +85,7 @@ bool RunTests()
     _clearfp();
     _controlfp_s(nullptr, fpcw, _MCW_EM);
 
-    print("Ran %d tests, %d pass, %d fail\n", nPass+nFail, nPass, nFail);
+    print("Ran %zu tests, %zu pass, %zu fail\n", nPass+nFail, nPass, nFail);
 
     return (nFail == 0);
 }
